+++ public/assets/data/products.json
@@ -1173,8 +1173,8 @@
             ],
 
             "img": [
-              "https://miraicollectibles.com/image/miraicollectibles/image/cache/data/all_product_images/product-1130/ofRbGmrs1658506743-1920x1080.jpeg",
-              "https://miraicollectibles.com/image/miraicollectibles/image/cache/data/all_product_images/product-1130/kdUC4wDX1658506738-1920x1080.jpeg"
+              "https://product.hstatic.net/200000515997/product/d1ebb03d-822d-4f7d-a1cf-3b84525fb598_003f5baf9b324c30b7f5d90fffd3edec_master.jpg",
+              "https://product.hstatic.net/200000515997/product/69cafbd3-4aab-423d-954a-187d86af0703_24137447dc6c473a80784ede3538cbfa_master.jpg"
             ],
             "nsfw": null
           },
@@ -1397,8 +1397,8 @@
               }
             ],
             "img": [
-              "https://img.orzgk.com/wp-content/uploads/2024/01/%E5%BE%AE%E4%BF%A1%E5%9B%BE%E7%89%87_20240103155302.jpg",
-              "https://www.senseitoy.com/image/magictoolbox_cache/cf3e6ec01aac7cb79461bcfe9d0d075e/2/0/2079_product/thumb600x600/2230335894/monkey-d-studio-luffy-red-roc-epic-diorama-resin-statue013.jpg"
+              "https://product.hstatic.net/200000515997/product/329ba6ed-4ee7-453b-8701-f6018720cbc5_ec5041127e514068ad6b6c759812b41b_master.jpg",
+              "https://product.hstatic.net/200000515997/product/ea08f9e4-8ded-4286-b444-f937eeaebb9b_49b4f25a0997494f99e03ccba68aff22_master.jpg"
             ],
             "nsfw": null
           }
