+++ src/App.js
@@ -4,7 +4,7 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Header from "./shared/components/Header/Header";
 import Home from "./features/home/Home";
 import Shop from "./features/shop/Shop";
-import License from "./features/license/License";
+import Licenses from "./features/licenses/Licenses";
 import Support from "./features/support/Support";
 import Info from "./features/info/Info";
 import Footer from "./shared/components/Footer/Footer";
@@ -12,6 +12,7 @@ import ScrollHandler from "./shared/hooks/ScrollHandler";
 import ScrollToTop from "./shared/hooks/ScrollToTop";
 import FAQs from "./features/faqs/FAQs";
 import ProductDetail from "./features/product/ProductDetail";
+import MustRead from "./shared/components/MustRead/MustRead";
 
 function App() {
   return (
@@ -23,10 +24,11 @@ function App() {
         <Route path="/" element={<Home />} />
         <Route path="/shop" element={<Shop />} />
         <Route path="/product/:id" element={<ProductDetail />} />
-        <Route path="/license" element={<License />} />
+        <Route path="/licenses" element={<Licenses />} />
         <Route path="/support" element={<Support />} />
         <Route path="/faqs" element={<FAQs />} />
         <Route path="/info" element={<Info />} />
+        <Route path="/must-read" element={<MustRead />} />
       </Routes>
       <Footer />
     </BrowserRouter>
