+++ src/features/shop/styles/Shop.css
@@ -68,6 +68,7 @@
   flex: 1;
   display: flex;
   flex-direction: column;
+  gap: 10px;
 }
 
 .sidebar-section:last-child {
